Move vendored package `rencode` from `synchronousdeluge` to `libs`
